build: maps feature

+++ app/src/main/java/com/dicoding/storyapp/data/source/remote/response/StoryResponse.kt
@@ -22,8 +22,8 @@ data class StoryResponse(
     val createdAt: String,
 
     @field:SerializedName("lat")
-    val lat: Float? = null,
+    val lat: Double? = null,
 
     @field:SerializedName("lon")
-    val lon: Float? = null
+    val lon: Double? = null
 ) : Parcelable
